feat: migrate sources.js
