+++ migration-to-ts/src/components/view/sources/sources.ts
@@ -1,0 +1,28 @@
+import { INewsItemSource } from '../../../types/types';
+import './sources.css';
+
+class Sources {
+  draw(data: INewsItemSource[]): void {
+    const fragment = document.createDocumentFragment();
+    const sourceItemTemp = document.querySelector('#sourceItemTemp') as HTMLTemplateElement;
+
+    data.forEach((item) => {
+      const sourceClone = sourceItemTemp.content.cloneNode(true) as Element | null;
+
+      const itemNameEl = sourceClone?.querySelector('.source__item-name') as HTMLTemplateElement;
+      itemNameEl.textContent = item.name;
+
+      const itemEl = sourceClone?.querySelector('.source__item') as HTMLTemplateElement;
+      itemEl.setAttribute('data-source-id', item.id);
+
+      if (sourceClone) {
+        fragment.append(sourceClone);
+      }
+    });
+
+    const sourcesEl = document.querySelector('.sources') as HTMLTemplateElement;
+    sourcesEl.append(fragment);
+  }
+}
+
+export default Sources;
